chore: reflect ibc-rs ADR010 changes
