+++ Cargo.toml
@@ -6,7 +6,7 @@ members = [
     "clients/sov-celestia-cw",
     "modules/sov-ibc-transfer",
     "modules/sov-ibc",
-    "mocks",
+    # "mocks",
 ]
 
 exclude = [
@@ -40,15 +40,15 @@ tempfile    = "3.5"
 thiserror   = "1.0.38"
 
 # ibc depedenencies
-ibc-core              = { git = "https://github.com/cosmos/ibc-rs.git", rev = "552863ec93", default-features = false, features = ["borsh","schema","serde"] }
-ibc-core-client       = { git = "https://github.com/cosmos/ibc-rs.git", rev = "552863ec93", default-features = false }
-ibc-core-host-cosmos  = { git = "https://github.com/cosmos/ibc-rs.git", rev = "552863ec93", default-features = false }
-ibc-client-tendermint = { git = "https://github.com/cosmos/ibc-rs.git", rev = "552863ec93", default-features = false }
-ibc-client-wasm-types = { git = "https://github.com/cosmos/ibc-rs.git", rev = "552863ec93", default-features = false }
-ibc-app-transfer      = { git = "https://github.com/cosmos/ibc-rs.git", rev = "552863ec93", default-features = false }
-ibc-primitives        = { git = "https://github.com/cosmos/ibc-rs.git", rev = "552863ec93", default-features = false }
-ibc-query             = { git = "https://github.com/cosmos/ibc-rs.git", rev = "552863ec93", default-features = false }
-ibc-testkit           = { git = "https://github.com/cosmos/ibc-rs.git", rev = "552863ec93", default-features = false }
+ibc-core              = { git = "https://github.com/cosmos/ibc-rs.git", rev = "4463366e65", default-features = false, features = ["borsh","schema","serde"] }
+ibc-core-client       = { git = "https://github.com/cosmos/ibc-rs.git", rev = "4463366e65", default-features = false }
+ibc-core-host-cosmos  = { git = "https://github.com/cosmos/ibc-rs.git", rev = "4463366e65", default-features = false }
+ibc-client-tendermint = { git = "https://github.com/cosmos/ibc-rs.git", rev = "4463366e65", default-features = false }
+ibc-client-wasm-types = { git = "https://github.com/cosmos/ibc-rs.git", rev = "4463366e65", default-features = false }
+ibc-app-transfer      = { git = "https://github.com/cosmos/ibc-rs.git", rev = "4463366e65", default-features = false }
+ibc-primitives        = { git = "https://github.com/cosmos/ibc-rs.git", rev = "4463366e65", default-features = false }
+ibc-query             = { git = "https://github.com/cosmos/ibc-rs.git", rev = "4463366e65", default-features = false }
+ibc-testkit           = { git = "https://github.com/cosmos/ibc-rs.git", rev = "4463366e65", default-features = false }
 ibc-proto             = { git = "https://github.com/cosmos/ibc-proto-rs.git", rev = "1b1d7a9", default-features = false }
 
 # cosmos dependencies
