+++ clients/sov-celestia-cw/src/context/mod.rs
@@ -1,5 +1,4 @@
 pub mod client_ctx;
-pub mod core_ctx;
 pub mod custom_ctx;
 
 use std::str::FromStr;
