+++ clients/sov-celestia-cw/src/handlers.rs
@@ -3,8 +3,8 @@ use ibc_core::client::context::client_state::{
     ClientStateCommon, ClientStateExecution, ClientStateValidation,
 };
 use ibc_core::client::context::consensus_state::ConsensusState as _;
+use ibc_core::client::context::ClientValidationContext;
 use ibc_core::host::types::path::ClientConsensusStatePath;
-use ibc_core::host::ValidationContext;
 use ibc_core::primitives::proto::Any;
 use prost::Message;
 use sov_celestia_client::client_state::ClientState;
