+++ clients/sov-celestia-cw/src/types/consensus_state.rs
@@ -2,19 +2,27 @@ use ibc_core::client::types::error::ClientError;
 use ibc_core::derive::ConsensusState as ConsensusStateDerive;
 use ibc_core::primitives::proto::{Any, Protobuf};
 use sov_celestia_client::consensus_state::ConsensusState;
-use sov_celestia_client::types::consensus_state::SOV_TENDERMINT_CONSENSUS_STATE_TYPE_URL;
+use sov_celestia_client::types::consensus_state::{
+    SovTmConsensusState, SOV_TENDERMINT_CONSENSUS_STATE_TYPE_URL,
+};
 
 #[derive(Clone, Debug, derive_more::From, ConsensusStateDerive)]
 pub enum AnyConsensusState {
     Sovereign(ConsensusState),
 }
 
-impl TryFrom<AnyConsensusState> for ConsensusState {
+impl From<SovTmConsensusState> for AnyConsensusState {
+    fn from(value: SovTmConsensusState) -> Self {
+        AnyConsensusState::Sovereign(value.into())
+    }
+}
+
+impl TryFrom<AnyConsensusState> for SovTmConsensusState {
     type Error = ClientError;
 
     fn try_from(value: AnyConsensusState) -> Result<Self, Self::Error> {
         match value {
-            AnyConsensusState::Sovereign(state) => Ok(state),
+            AnyConsensusState::Sovereign(state) => Ok(state.inner().clone()),
         }
     }
 }
