+++ clients/sov-celestia/src/context.rs
@@ -1,60 +1,67 @@
-use alloc::string::ToString;
-
-use ibc_core::client::context::ClientExecutionContext;
+use ibc_core::client::context::{ClientExecutionContext, ClientValidationContext};
+use ibc_core::client::types::error::ClientError;
 use ibc_core::client::types::Height;
 use ibc_core::handler::types::error::ContextError;
 use ibc_core::host::types::identifiers::ClientId;
-use ibc_core::host::types::path::ClientConsensusStatePath;
 use ibc_core::primitives::prelude::*;
 use ibc_core::primitives::Timestamp;
+use sov_celestia_client_types::consensus_state::SovTmConsensusState;
 
-use crate::consensus_state::ConsensusState;
+/// Enables conversion (`TryInto` and `From`) between the consensus state type
+/// used by the host and the one specific to the Tendermint light client, which
+/// is `ConsensusStateType`.
+pub trait ConsensusStateConverter:
+    TryInto<SovTmConsensusState, Error = ClientError> + From<SovTmConsensusState>
+{
+}
 
-/// Client's context required during both validation and execution
-pub trait CommonContext {
-    type ConversionError: ToString;
-    type AnyConsensusState: TryInto<ConsensusState, Error = Self::ConversionError>;
+impl<C> ConsensusStateConverter for C where
+    C: TryInto<SovTmConsensusState, Error = ClientError> + From<SovTmConsensusState>
+{
+}
 
+/// Client's context required during validation
+pub trait ValidationContext: ClientValidationContext
+where
+    Self::ConsensusStateRef: ConsensusStateConverter,
+{
     /// Returns the current timestamp of the local chain.
     fn host_timestamp(&self) -> Result<Timestamp, ContextError>;
 
     /// Returns the current height of the local chain.
     fn host_height(&self) -> Result<Height, ContextError>;
 
-    /// Retrieve the consensus state for the given client ID at the specified
-    /// height.
-    ///
-    /// Returns an error if no such state exists.
-    fn consensus_state(
-        &self,
-        client_cons_state_path: &ClientConsensusStatePath,
-    ) -> Result<Self::AnyConsensusState, ContextError>;
-
     /// Returns all the heights at which a consensus state is stored
     fn consensus_state_heights(&self, client_id: &ClientId) -> Result<Vec<Height>, ContextError>;
-}
 
-/// Client's context required during validation
-pub trait ValidationContext: CommonContext {
     /// Search for the lowest consensus state higher than `height`.
     fn next_consensus_state(
         &self,
         client_id: &ClientId,
         height: &Height,
-    ) -> Result<Option<Self::AnyConsensusState>, ContextError>;
+    ) -> Result<Option<Self::ConsensusStateRef>, ContextError>;
 
     /// Search for the highest consensus state lower than `height`.
     fn prev_consensus_state(
         &self,
         client_id: &ClientId,
         height: &Height,
-    ) -> Result<Option<Self::AnyConsensusState>, ContextError>;
+    ) -> Result<Option<Self::ConsensusStateRef>, ContextError>;
 }
 
 /// Client's context required during execution.
 ///
 /// This trait is automatically implemented for all types that implement
-/// [`CommonContext`] and [`ClientExecutionContext`]
-pub trait ExecutionContext: CommonContext + ClientExecutionContext {}
+/// [`ValidationContext`] and [`ClientExecutionContext`]
+pub trait ExecutionContext: ValidationContext + ClientExecutionContext
+where
+    Self::ConsensusStateRef: ConsensusStateConverter,
+{
+}
 
-impl<T> ExecutionContext for T where T: CommonContext + ClientExecutionContext {}
+impl<T> ExecutionContext for T
+where
+    T: ValidationContext + ClientExecutionContext,
+    T::ConsensusStateRef: ConsensusStateConverter,
+{
+}
