+++ clients/sov-celestia/types/src/consensus_state/definition.rs
@@ -5,6 +5,7 @@ use ibc_core::commitment_types::commitment::CommitmentRoot;
 use ibc_core::primitives::proto::{Any, Protobuf};
 use ibc_proto::ibc::lightclients::sovereign::tendermint::v1::ConsensusState as RawConsensusState;
 use prost::Message;
+use tendermint::Time;
 
 use super::TmConsensusParams;
 use crate::client_message::SovTmHeader;
@@ -28,6 +29,10 @@ impl<Da> ConsensusState<Da> {
 pub type SovTmConsensusState = ConsensusState<TmConsensusParams>;
 
 impl SovTmConsensusState {
+    pub fn timestamp(&self) -> Time {
+        self.da_params.timestamp
+    }
+
     /// Protobuf encoding of the `SovTmConsensusState` through the `Any` type.
     pub fn encode_thru_any(self) -> Vec<u8> {
         Any::from(self).encode_to_vec()
