+++ mocks/Cargo.toml
@@ -38,7 +38,7 @@ ibc-app-transfer      = { workspace = true }
 ibc-client-tendermint = { workspace = true }
 ibc-core-host-cosmos  = { workspace = true }
 ibc-query             = { workspace = true }
-ibc-testkit           = { git = "https://github.com/cosmos/ibc-rs.git", rev = "552863ec93", default-features = false }
+ibc-testkit           = { git = "https://github.com/cosmos/ibc-rs.git", rev = "a9761140d3", default-features = false }
 
 # cosmos dependencies
 basecoin-app         = { git = "https://github.com/informalsystems/basecoin-rs.git", rev = "e668ec4" }
