+++ modules/sov-ibc/src/clients/mod.rs
@@ -4,6 +4,7 @@ use derive_more::{From, TryInto};
 use ibc_client_tendermint::client_state::ClientState as TmClientState;
 use ibc_client_tendermint::consensus_state::ConsensusState as TmConsensusState;
 use ibc_client_tendermint::types::{
+    ClientState as TmClientStateTypes, ConsensusState as TmConsensusStateType,
     TENDERMINT_CLIENT_STATE_TYPE_URL, TENDERMINT_CONSENSUS_STATE_TYPE_URL,
 };
 use ibc_core::client::context::client_state::{
@@ -20,8 +21,12 @@ use ibc_core::host::types::path::Path;
 use ibc_core::primitives::proto::{Any, Protobuf};
 use sov_celestia_client::client_state::ClientState as SovClientState;
 use sov_celestia_client::consensus_state::ConsensusState as SovConsensusState;
-use sov_celestia_client::types::client_state::SOV_TENDERMINT_CLIENT_STATE_TYPE_URL;
-use sov_celestia_client::types::consensus_state::SOV_TENDERMINT_CONSENSUS_STATE_TYPE_URL;
+use sov_celestia_client::types::client_state::{
+    SovTmClientState, SOV_TENDERMINT_CLIENT_STATE_TYPE_URL,
+};
+use sov_celestia_client::types::consensus_state::{
+    SovTmConsensusState, SOV_TENDERMINT_CONSENSUS_STATE_TYPE_URL,
+};
 use sov_modules_api::{DaSpec, Spec};
 
 use crate::context::IbcContext;
@@ -30,6 +35,18 @@ use crate::context::IbcContext;
 pub enum AnyClientState {
     Tendermint(TmClientState),
     Sovereign(SovClientState),
+}
+
+impl From<SovTmClientState> for AnyClientState {
+    fn from(cs: SovTmClientState) -> Self {
+        Self::Sovereign(cs.into())
+    }
+}
+
+impl From<TmClientStateTypes> for AnyClientState {
+    fn from(cs: TmClientStateTypes) -> Self {
+        Self::Tendermint(cs.into())
+    }
 }
 
 impl Protobuf<Any> for AnyClientState {}
@@ -269,10 +286,48 @@ where
     }
 }
 
-#[derive(Clone, From, TryInto, ConsensusState)]
+#[derive(Clone, From, ConsensusState)]
 pub enum AnyConsensusState {
     Tendermint(TmConsensusState),
     Sovereign(SovConsensusState),
+}
+
+impl TryFrom<AnyConsensusState> for TmConsensusStateType {
+    type Error = ClientError;
+
+    fn try_from(cs: AnyConsensusState) -> Result<TmConsensusStateType, Self::Error> {
+        match cs {
+            AnyConsensusState::Tendermint(cs) => Ok(cs.inner().clone()),
+            _ => Err(ClientError::UnknownConsensusStateType {
+                consensus_state_type: "".to_string(),
+            }),
+        }
+    }
+}
+
+impl TryFrom<AnyConsensusState> for SovTmConsensusState {
+    type Error = ClientError;
+
+    fn try_from(cs: AnyConsensusState) -> Result<SovTmConsensusState, Self::Error> {
+        match cs {
+            AnyConsensusState::Sovereign(cs) => Ok(cs.inner().clone()),
+            _ => Err(ClientError::UnknownConsensusStateType {
+                consensus_state_type: "".to_string(),
+            }),
+        }
+    }
+}
+
+impl From<SovTmConsensusState> for AnyConsensusState {
+    fn from(cs: SovTmConsensusState) -> Self {
+        Self::Sovereign(cs.into())
+    }
+}
+
+impl From<TmConsensusStateType> for AnyConsensusState {
+    fn from(cs: TmConsensusStateType) -> Self {
+        Self::Tendermint(cs.into())
+    }
 }
 
 impl Protobuf<Any> for AnyConsensusState {}
